Work advances: separate claim's functions from rrDisplay
