+++ src/app/Score.ts
@@ -1,12 +1,9 @@
 import Root from './Root';
 import Claim from './Claim';
-import Operation from './Operation';
-
 
 ﻿export default class Score {
     claimId: string;
     selectedScore: Score;
-    operation: Operation = new Operation();
     /**  */
     confidencePro: number;
     confidenceCon: number;
@@ -56,13 +53,6 @@ import Operation from './Operation';
     isMain: boolean;
     isEditing: boolean;
 
-    select(score: Score, e: Event): void {
-        if (score != this.selectedScore) {
-            this.selectedScore = score;
-            this.operation.setDisplayState();
-            this.operation.update();
-        }
-    }
 }
 
 type DisplayState = "newClaim" | "notSelected" | "parent" | "ancestor" | "selected"| "selected editing" | "child" | "sibling";
